refactor: Compose 공통으로 사용하는 부분들 core 모듈로 이동
