+++ features/compose-ui/src/main/java/com/feature/compose_ui/ComposeUiActivity.kt
@@ -16,6 +16,7 @@ import androidx.compose.ui.platform.ViewCompositionStrategy
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.feature.compose_ui.databinding.AComposeUiBinding
+import com.hmju.core.compose.TilTheme
 import com.hmju.core.ui.base.BaseActivity
 import timber.log.Timber
