+++ features/compose-ui/src/main/java/com/feature/compose_ui/GeneralComposeActivity.kt
@@ -19,6 +19,7 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
+import com.hmju.core.compose.TilTheme
 import timber.log.Timber
 
 /**
