+++ features/compose-ui/src/main/java/com/feature/compose_ui/MemoComposeUiActivity.kt
@@ -30,6 +30,7 @@ import androidx.lifecycle.lifecycleScope
 import com.feature.compose_ui.model.MemoModel
 import com.feature.compose_ui.model.MemoUiModel
 import com.feature.compose_ui.usecase.GetMemoListUseCase
+import com.hmju.core.compose.TilTheme
 import com.hmju.core.model.params.PagingParameter
 import com.hmju.core.ui.paging.PagingModel
 import dagger.hilt.android.AndroidEntryPoint
